Сервисному слою добавлены интерфейсы
